+++ rollup.config.js
@@ -1,0 +1,27 @@
+import resolve from 'rollup-plugin-node-resolve';
+import babel from 'rollup-plugin-babel';
+import commonjs from 'rollup-plugin-commonjs';
+
+export default {
+  input: 'src/index.js',
+  output: {
+    file: 'dist/bundle.js',
+    format: 'cjs',
+    name: 'ObservePropsDecorator',
+    sourcemap: true
+  },
+  plugins: [
+    resolve({
+      customResolveOptions: {
+        moduleDirectory: 'node_modules'
+      }
+    }),
+    commonjs({
+      include: 'node_modules/**'
+    }),
+    babel({
+      exclude: 'node_modules/**'
+    })
+  ],
+  external: ['react', 'react-dom', 'rxjs']
+};